apps/backend/requirements.txt
# Core Framework
fastapi==0.124.4
uvicorn[standard]==0.34.0
pydantic==2.12.5
pydantic-settings==2.7.1
pydantic[email]==2.12.5

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==25.1.0  # Required for encryption key derivation
python-multipart==0.0.12

# Rate Limiting
slowapi==0.1.9

# Email & SMS
resend==2.4.0
africastalking==1.2.6
twilio==9.3.4

# Encryption & File Storage
cryptography==43.0.3
cloudinary==1.40.0
boto3==1.35.47

# PDF Generation
weasyprint==62.3
jinja2==3.1.4

# HTTP & API
requests==2.32.3
httpx==0.27.2

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0

# Observability
sentry-sdk[fastapi]==2.20.0
prometheus-client==0.21.0

# Utilities
python-dotenv==1.0.1
